backend/requirements.txt
tensorflow-cpu
numpy
Pillow
flask
flask-cors
scikit-learn
gunicorn
